Breaking out scriptSkipping, smartStep, and stackFrames
